moved table to separate file
